backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase>=2.0.0
pydantic==2.9.2
pydantic-settings==2.5.2
apscheduler==3.10.4
google-generativeai==0.8.3
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
websockets==13.1
google-api-python-client==2.149.0
google-auth-oauthlib==1.2.1
tzdata==2024.2
sentry-sdk[fastapi]>=2.0.0
slowapi>=0.1.9
email-validator>=2.0.0
litellm>=1.40.0
langfuse>=3.7.0
tenacity>=8.0.0
